Fix read overflow on arrays
The arrays errstr_table and sequences_table are declared as [][2].
The compiler is free to allocate as much as the highest number of
index requires, in this case 126.

The function they passed to expects them to be [128][2] and GCC 11
complains that there is actually an overflow.

Declare these arrays explicitly to be [128][2].

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>