Create a libamx shared library and install it
Create a libamx shared library and install it along with other
dlopened modules and executables. Also add a pkgconfig file.

Link amxString with curses if it was built with it.
Don't expect that the executable dlopen()ing amxString has been
linked with curses. This will fix missing symbols shown by
ldd -r -d amxString.so

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>